requirement.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
APScheduler==3.6.3
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
bs4==0.0.1
collection==0.1.6
dnspython==2.3.0
frozenlist==1.3.3
h11==0.14.0
html5lib==1.1
httpcore==0.16.3
httpx==0.23.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
pyaes==1.6.1
pymongo==4.3.2
Pyrogram==2.0.106
PySocks==1.7.1
pyTelegramBotAPI==4.11.0
python-telegram-bot==13.13
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.29.0
rfc3986==1.5.0
urllib3==1.26.15